# pkg/api/doc_ls.go
/*
Copyright © 2020 FairOS Authors

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

package api

import (
	"net/http"

	"github.com/fairdatasociety/fairOS-dfs/pkg/collection"
	"github.com/fairdatasociety/fairOS-dfs/pkg/cookie"
	"resenje.org/jsonhttp"
)

// DocumentDBs represent a list of documentDB
type DocumentDBs struct {
	Tables []documentDB
}

type documentDB struct {
	Name           string              `json:"tableName"`
	IndexedColumns []collection.SIndex `json:"indexes"`
	CollectionType string              `json:"type"`
}

// DocListHandler godoc
//
//	@Summary      List all doc table
//	@Description  DocListHandler is the api handler which lists all the document database in a pod
//	@Tags         doc
//	@Accept       json
//	@Produce      json
//	@Param	      podName query string true "pod name"
//	@Param	      Cookie header string true "cookie parameter"
//	@Success      200  {object}  DocumentDBs
//	@Failure      400  {object}  response
//	@Failure      500  {object}  response
//	@Router       /v1/doc/ls [get]
func (h *Handler) DocListHandler(w http.ResponseWriter, r *http.Request) {
	keys, ok := r.URL.Query()["podName"]
	if !ok || len(keys[0]) < 1 {
		h.logger.Errorf("doc ls: \"podName\" argument missing")
		jsonhttp.BadRequest(w, &response{Message: "doc ls: \"podName\" argument missing"})
		return
	}
	podName := keys[0]
	if podName == "" {
		h.logger.Errorf("doc ls: \"podName\" argument missing")
		jsonhttp.BadRequest(w, &response{Message: "doc ls: \"podName\" argument missing"})
		return
	}

	// get values from cookie
	sessionId, err := cookie.GetSessionIdFromCookie(r)
	if err != nil {
		h.logger.Errorf("doc ls: invalid cookie: %v", err)
		jsonhttp.BadRequest(w, &response{Message: ErrInvalidCookie.Error()})
		return
	}
	if sessionId == "" {
		h.logger.Errorf("doc ls: \"cookie-id\" parameter missing in cookie")
		jsonhttp.BadRequest(w, &response{Message: "doc ls: \"cookie-id\" parameter missing in cookie"})
		return
	}

	collections, err := h.dfsAPI.DocList(sessionId, podName)
	if err != nil {
		h.logger.Errorf("doc ls: %v", err)
		jsonhttp.InternalServerError(w, &response{Message: "doc ls: " + err.Error()})
		return
	}

	var col DocumentDBs
	for name, dbSchema := range collections {
		var indexes []collection.SIndex
		indexes = append(indexes, dbSchema.SimpleIndexes...)
		indexes = append(indexes, dbSchema.MapIndexes...)
		indexes = append(indexes, dbSchema.ListIndexes...)
		m := documentDB{
			Name:           name,
			IndexedColumns: indexes,
			CollectionType: "Document Store",
		}
		col.Tables = append(col.Tables, m)
	}

	jsonhttp.OK(w, col)
}