# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A list of the currently provided Interfaces to interact with FairOS decentralised file system(dfs), implementing user, pod, file system, key value store and document store",
        "title": "FairOS-dfs server",
        "contact": {
            "name": "Sabyasachi Patra",
            "email": "sabyasachi@datafund.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.0.0"
    },
    "paths": {
...
        "/v1/dir/rmdir": {
            "delete": {
                "description": "DirectoryRmdirHandler is the api handler to remove a directory.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dir"
                ],
                "summary": "Remove directory",
                "parameters": [
                    {
                        "description": "pod name and dir path",
                        "name": "dir_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DirRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
        "/v1/doc/delete": {
            "delete": {
                "description": "DocDeleteHandler is the api handler to delete the given document database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doc"
                ],
                "summary": "Delete a doc table",
                "parameters": [
                    {
                        "description": "doc table info",
                        "name": "doc_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SimpleDocRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
        "/v1/doc/ls": {
            "get": {
                "description": "DocListHandler is the api handler which lists all the document database in a pod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doc"
                ],
                "summary": "List all doc table",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pod name",
                        "name": "podName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DocumentDBs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
        "/v1/doc/open": {
            "post": {
                "description": "DocOpenHandler is the api handler to open a document database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "doc"
                ],
                "summary": "Open a doc table",
                "parameters": [
                    {
                        "description": "doc table info",
                        "name": "doc_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DocRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DocumentDBs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.DirRequest": {
            "type": "object",
            "properties": {
                "dirPath": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                }
            }
        },
...
        "api.DocRequest": {
            "type": "object",
            "properties": {
                "mutable": {
                    "type": "boolean"
                },
                "podName": {
                    "type": "string"
                },
                "si": {
                    "type": "string"
                },
                "tableName": {
                    "type": "string"
                }
            }
        },
        "api.DocumentDBs": {
            "type": "object",
            "properties": {
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.documentDB"
                    }
                }
            }
        },
...
        "api.SimpleDocRequest": {
            "type": "object",
            "properties": {
                "podName": {
                    "type": "string"
                },
                "tableName": {
                    "type": "string"
                }
            }
        },
...
        "api.documentDB": {
            "type": "object",
            "properties": {
                "indexes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/collection.SIndex"
                    }
                },
                "tableName": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "collection.SIndex": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
    }
}